+++ main-server/src/core/common/thai_citizen_id.js
@@ -1,0 +1,42 @@
+/**
+ * Copyright (c) 2018, 2019 National Digital ID COMPANY LIMITED
+ *
+ * This file is part of NDID software.
+ *
+ * NDID is the free software: you can redistribute it and/or modify it under
+ * the terms of the Affero GNU General Public License as published by the
+ * Free Software Foundation, either version 3 of the License, or any later
+ * version.
+ *
+ * NDID is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the Affero GNU General Public License for more details.
+ *
+ * You should have received a copy of the Affero GNU General Public License
+ * along with the NDID source code. If not, see https://www.gnu.org/licenses/agpl.txt.
+ *
+ * Please contact info@ndid.co.th for any further questions
+ *
+ */
+
+import * as thaiIdUtils from '../../utils/thai_id';
+import CustomError from 'ndid-error/custom_error';
+import errorType from 'ndid-error/type';
+
+const CITIZEN_ID_NAMESPACE = 'citizen_id';
+
+export function validateIdentifier({ namespace, identifier }) {
+  if (namespace === CITIZEN_ID_NAMESPACE) {
+    const valid = thaiIdUtils.validateThaiIdNumber(identifier);
+    if (!valid) {
+      throw new CustomError({
+        errorType: errorType.INVALID_IDENTIFIER_KEY,
+        details: {
+          namespace,
+          identifier,
+        },
+      });
+    }
+  }
+}